Renamed perf.py to profile_util.py
